integration/UserTest -> teste de integração com mockbean -> findAllUsers busca todos os users, postUser cria um user e procura o mesmo user especifico.
